Reorganize trinity modules
