+++ tests/trinity/core/initialization/test_is_data_dir_initialized.py
@@ -3,7 +3,7 @@ import pytest
 
 from eth_utils import decode_hex
 
-from trinity.chains import (
+from trinity.initialization import (
     is_data_dir_initialized,
 )
 from trinity.config import (
