+++ tests/trinity/core/initialization/test_is_database_initialized.py
@@ -5,7 +5,7 @@ from eth.chains.ropsten import ROPSTEN_GENESIS_HEADER
 from eth.db.backends.level import LevelDB
 from eth.db.chain import ChainDB
 
-from trinity.chains import (
+from trinity.initialization import (
     initialize_data_dir,
     is_database_initialized,
 )
