+++ trinity/protocol/common/constants.py
@@ -1,4 +1,3 @@
-
 # The default timeout for a round trip API request and response from a peer.
 ROUND_TRIP_TIMEOUT = 20.0
