--- trinity/utils/db_proxy.py
@@ -1,34 +1,0 @@
-from multiprocessing.managers import (
-    BaseManager,
-)
-import pathlib
-
-from trinity.chains import (
-    AsyncHeaderChainProxy,
-    ChainProxy,
-)
-from trinity.db.chain import ChainDBProxy
-from trinity.db.base import DBProxy
-from trinity.db.header import (
-    AsyncHeaderDBProxy
-)
-
-
-def create_db_manager(ipc_path: pathlib.Path) -> BaseManager:
-    """
-    We're still using 'str' here on param ipc_path because an issue with
-    multi-processing not being able to interpret 'Path' objects correctly
-    """
-    class DBManager(BaseManager):
-        pass
-
-    # Typeshed definitions for multiprocessing.managers is incomplete, so ignore them for now:
-    # https://github.com/python/typeshed/blob/85a788dbcaa5e9e9a62e55f15d44530cd28ba830/stdlib/3/multiprocessing/managers.pyi#L3
-    DBManager.register('get_db', proxytype=DBProxy)  # type: ignore
-    DBManager.register('get_chaindb', proxytype=ChainDBProxy)  # type: ignore
-    DBManager.register('get_chain', proxytype=ChainProxy)  # type: ignore
-    DBManager.register('get_headerdb', proxytype=AsyncHeaderDBProxy)  # type: ignore
-    DBManager.register('get_header_chain', proxytype=AsyncHeaderChainProxy)  # type: ignore
-
-    manager = DBManager(address=str(ipc_path))  # type: ignore
-    return manager
