+++ trinity/utils/logging.py
@@ -10,8 +10,8 @@ from logging.handlers import (
     QueueHandler,
     RotatingFileHandler,
 )
-
 import os
+from pathlib import Path
 import sys
 from typing import (
     Any,
@@ -26,10 +26,6 @@ from cytoolz import dissoc
 
 from eth.tools.logging import (
     TraceLogger,
-)
-
-from trinity.config import (
-    TrinityConfig,
 )
 
 if TYPE_CHECKING:
@@ -97,7 +93,7 @@ def setup_trinity_file_and_queue_logging(
         logger: Logger,
         formatter: Formatter,
         handler_stream: StreamHandler,
-        trinity_config: TrinityConfig,
+        logfile_path: Path,
         level: int=None) -> Tuple[Logger, 'Queue[str]', QueueListener]:
     from .mp import ctx
 
@@ -107,7 +103,7 @@ def setup_trinity_file_and_queue_logging(
     log_queue = ctx.Queue()
 
     handler_file = RotatingFileHandler(
-        str(trinity_config.logfile_path),
+        str(logfile_path),
         maxBytes=(10000000 * LOG_MAX_MB),
         backupCount=LOG_BACKUP_COUNT
     )
